+++ src/main/java/org/thinkingstudio/bedsheet/BedSheetModEntrypoint.java
@@ -3,18 +3,21 @@ package org.thinkingstudio.bedsheet;
 import carpet.CarpetServer;
 import carpet.utils.CarpetRulePrinter;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.loading.FMLLoader;
 import net.neoforged.neoforge.common.NeoForge;
+import org.thinkingstudio.bedsheet.loader.entrypoint.DedicatedServerModInitializer;
+import org.thinkingstudio.bedsheet.loader.entrypoint.EntrypointHandler;
 
 @Mod(BedSheetModReference.MODID)
 public class BedSheetModEntrypoint {
-
-    public BedSheetModEntrypoint(IEventBus modEventBus) {
+    public BedSheetModEntrypoint(ModContainer modContainer, IEventBus modEventBus) {
+        EntrypointHandler.init();
         CarpetServer.onGameStarted();
         BedSheetModEvents.registerEvents(modEventBus, NeoForge.EVENT_BUS);
         if (FMLLoader.getDist().isDedicatedServer()) {
-            CarpetRulePrinter.onInitializeServer();
+            modContainer.registerExtensionPoint(DedicatedServerModInitializer.class, new CarpetRulePrinter());
         }
     }
 }