+++ src/main/java/org/thinkingstudio/bedsheet/loader/FoxifiedLoader.java
@@ -1,4 +1,4 @@
-package org.thinkingstudio.bedsheet.util;
+package org.thinkingstudio.bedsheet.loader;
 
 import cpw.mods.modlauncher.ArgumentHandler;
 import cpw.mods.modlauncher.Launcher;
@@ -14,7 +14,7 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.Optional;
 
-public class NeoHooks {
+public class FoxifiedLoader {
     private static String[] launchArgs;
 
     public static boolean isDevelopmentEnvironment() {