+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Changelog
-- fix compat with AnvilCraft's `PistonBaseBlockMixin` (#7)
+- change: `org/thinkingstudio/bedsheet/util/NeoHooks.java` -> `org/thinkingstudio/bedsheet/loader/FoxifiedLoader.java`